refactor(spec): restore strict architecture boundaries
